Adding support for localAuthentication
