+++ SampleApps/PingExample/PingExample/JourneyView.swift
@@ -141,7 +141,6 @@ struct JourneyNodeView: View {
             callback is SuspendedTextOutputCallback ||
             callback is PingOneProtectInitializeCallback ||
             callback is PingOneProtectEvaluationCallback ||
-            callback is SelectIdpCallback ||
             callback is IdpCallback
         }
     }
@@ -218,7 +217,14 @@ struct JourneyNodeView: View {
             }
             
             if showNext {
-                Button(action: { onNext() }) {
+                Button(action: {
+                    if let selectIDPCallback = continueNode.callbacks.first(where: {
+                        $0 is SelectIdpCallback
+                    }) as? SelectIdpCallback {
+                        selectIDPCallback.value = "localAuthentication"
+                    }
+                    onNext()
+                }) {
                     Text("Next")
                         .frame(maxWidth: .infinity)
                         .padding()
@@ -227,131 +233,6 @@ struct JourneyNodeView: View {
                         .cornerRadius(8)
                 }
                 .padding(.top, 16)
-            }
-        }
-        .padding()
-    }
-}
-
-struct IdpCallbackView: View {
-    @StateObject var viewModel: IdpCallbackViewModel
-    
-    let onNext: () -> Void
-    
-    var body: some View {
-        VStack(spacing: 20) {
-            switch viewModel.authState {
-            case .authenticating:
-                Image(systemName: "person.circle.fill")
-                    .font(.largeTitle)
-                    .foregroundColor(.white)
-                Text(viewModel.callback.provider)
-                    .font(.body)
-                    .foregroundColor(.secondary)
-                    NextButton(title: "Continue") {
-                        Task { @MainActor in
-                            if viewModel.hasStartedAuthorization == false {
-                                // Call the performAuthorization method to start the process.
-                                await viewModel.performAuthorization()
-                            }
-                        }
-                    }
-            case .failure(let error):
-                Image(systemName: "xmark.octagon.fill")
-                    .font(.largeTitle)
-                    .foregroundColor(.red)
-                Text("Authorization failed: \(error.localizedDescription)")
-                    .font(.body)
-                    .multilineTextAlignment(.center)
-                NextButton(title: "Continue") {
-                    Task {
-                        self.onNext()
-                    }
-                }
-                .buttonStyle(.bordered)
-                
-            case .completed:
-                Image(systemName: "checkmark.circle.fill")
-                    .font(.largeTitle)
-                    .foregroundColor(.green)
-                Text("Authorization Successful")
-                    .font(.body)
-                    .multilineTextAlignment(.center)
-                NextButton(title: "Continue") {
-                    Task {
-                        self.onNext()
-                    }
-                }
-            }
-        }
-        .padding()
-    }
-}
-
-@MainActor
-class IdpCallbackViewModel: ObservableObject {
-    
-    @Published var authState: AuthState = .authenticating
-    
-    // 1. Add a flag to track if the task has been started.
-    var hasStartedAuthorization = false
-    
-    let callback: IdpCallback
-
-    enum AuthState {
-        case authenticating
-        case completed
-        case failure(Error)
-    }
-    
-    init(callback: IdpCallback) {
-        self.callback = callback
-    }
-    
-    func performAuthorization() async {
-        Task { @MainActor in
-            // 2. Check the flag. If the task has already run, do nothing.
-            guard !hasStartedAuthorization else { return }
-            hasStartedAuthorization = true
-            
-            let result = await callback.authorize()
-            
-            switch result {
-            case .success:
-                self.authState = .completed
-            case .failure(let error):
-                self.authState = .failure(error)
-            }
-        }
-    }
-}
-
-struct SelectIdpCallbackView: View {
-    let callback: SelectIdpCallback
-    let onNext: () -> Void
-    
-    var body: some View {
-        ScrollView {
-            
-            LazyVStack(alignment: .center, spacing: 12) {
-                
-                // Add a title for better context
-                Text("Select a provider")
-                    .font(.headline)
-                    .padding(.bottom, 8)
-                
-                ForEach(callback.providers) { provider in
-                    Button(action: {
-                        callback.value = provider.provider
-                        self.onNext()
-                    }) {
-                        // Make the button label more descriptive and visually appealing
-                        Text(provider.provider.capitalized)
-                            .fontWeight(.semibold)
-                            .frame(maxWidth: .infinity)
-                    }
-                    .buttonStyle(.borderedProminent)
-                }
             }
         }
         .padding()
